+++ truffle/src/com.oracle.truffle.dsl.processor/src/com/oracle/truffle/dsl/processor/verify/VerifyTruffleProcessor.java
@@ -146,9 +146,11 @@ public class VerifyTruffleProcessor extends AbstractProcessor {
                     for (VariableElement parameter : method.getParameters()) {
                         Element paramType = processingEnv.getTypeUtils().asElement(parameter.asType());
                         if (paramType != null && (paramType.equals(virtualFrameType) || paramType.equals(frameType))) {
-                            errorMessage(element, "Method %s cannot be annotated with @%s and have a parameter of type %s", method.getSimpleName(),
-                                            types.CompilerDirectives_TruffleBoundary.asElement().getSimpleName().toString(),
-                                            paramType.getSimpleName());
+                            CharSequence truffleBoundarySimpleName = types.CompilerDirectives_TruffleBoundary.asElement().getSimpleName();
+                            errorMessage(element, "Method %s cannot be annotated with @%s and have a parameter of type %s.%n" +
+                                            "To resolve this, either change the parameter to a %s, remove the parameter or remove the @%s.", method.getSimpleName(),
+                                            truffleBoundarySimpleName, paramType.getSimpleName(), types.MaterializedFrame.asElement().getSimpleName(),
+                                            truffleBoundarySimpleName);
                         }
                     }
                 } catch (Throwable t) {