+++ truffle/CHANGELOG.md
@@ -72,6 +72,7 @@ This changelog summarizes major changes between Truffle versions relevant to lan
 * GR-34894 Introduced `Location.isPrimitive()`, `Location.getConstantValue()`, and `Shape.makePropertyGetter(Object)`.
 * GR-39058 The Static Object Model offers preliminary support for field-based storage also on Native Image. 
 * GR-24927 A compilation error is emitted for methods annotated by a `@TruffleBoundary` annotation and taking a `Frame` parameter.
+* GR-24927 The Truffle annotation processor now emits an error for methods annotated by a `@TruffleBoundary` annotation and a `Frame` parameter. Previously, the processor only reported an error for `VirtualFrame` parameters. To resolve this, either change the parameter to a `MaterializedFrame` , remove the parameter or remove the `@TruffleBoundary`.
 
 ## Version 22.1.0